move smart contract context
